+++ config/applications/web-guitar-pro/bundles.php
@@ -9,7 +9,6 @@ return [
     Symfony\Bundle\TwigBundle\TwigBundle::class => ['all' => true],
     Symfony\Bundle\DebugBundle\DebugBundle::class => ['dev' => true, 'test' => true],
 	Symfony\WebpackEncoreBundle\WebpackEncoreBundle::class => ['all' => true],
-    //Symfony\Bundle\SwiftmailerBundle\SwiftmailerBundle::class => ['all' => true],
 	
 	SymfonyCasts\Bundle\VerifyEmail\SymfonyCastsVerifyEmailBundle::class => ['all' => true],
     SymfonyCasts\Bundle\ResetPassword\SymfonyCastsResetPasswordBundle::class => ['all' => true],