+++ themes/WebGuitarPro_AngularJs/assets/js/Player/index.js
@@ -1,0 +1,4 @@
+//require( './alphatab.js' );
+
+import './polyfills.ts';
+import './main.ts';