+++ themes/WebGuitarPro_AngularJs/assets/js/Player/tablature-player/player-controls/play-pause-button-item/play-pause-button-item.component.ts
@@ -1,12 +1,18 @@
 import { Component, Input, OnInit } from '@angular/core';
 import { AlphaTabApi } from '@coderline/alphatab';
 
+import templateString from './play-pause-button-item.component.html'
+
 declare var $: any;
 
 @Component({
     selector: 'play-pause-button-item',
-    templateUrl: './play-pause-button-item.component.html',
-    styleUrls: ['../player-controls.component.scss']
+    
+    template: templateString || 'Template Not Loaded !!!',
+    //templateUrl: './play-pause-button-item.component.html',
+    
+    styleUrls: []
+    //styleUrls: ['../player-controls.component.scss']
 })
 export class PlayPauseButtonItemComponent implements OnInit
 {