+++ themes/WebGuitarPro_AngularJs/assets/js/Player/tablature-player/player-controls/song-position-item/song-position-item.component.ts
@@ -1,10 +1,16 @@
 import { Component, Input, OnInit } from '@angular/core';
 import { AlphaTabApi } from '@coderline/alphatab';
 
+import templateString from './song-position-item.component.html'
+
 @Component({
     selector: 'song-position-item',
-    templateUrl: './song-position-item.component.html',
-    styleUrls: ['../player-controls.component.scss']
+    
+    template: templateString || 'Template Not Loaded !!!',
+    //templateUrl: './song-position-item.component.html',
+    
+    styleUrls: []
+    //styleUrls: ['../player-controls.component.scss']
 })
 export class SongPositionItemComponent implements OnInit
 {